# openapi.yaml
openapi: 3.1.0
info:
  title: Food Bank API
  version: '1.0.0'
  description: |
    JSON API for running a food bank: referrals in, sessions scheduled, pick
    lists produced, stock adjusted.

    **Read `API.md` alongside this.** A schema describes shapes; it cannot
    describe the login/refresh cycle, what happens to a referral from an
    unrecognised address, or why a team lead receives fewer fields than an
    admin. Those are the parts most likely to be got wrong.

    ## Generating a client

    ```
    npx openapi-typescript openapi.yaml -o src/api/schema.d.ts
    ```
servers:
  - url: http://127.0.0.1:8787
    description: Local wrangler dev
  - url: https://foodbank-server-production.workers.dev
    description: Production
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/sessions/{sessionId}/listener-sheet:
    parameters: [{ $ref: '#/components/parameters/SessionId' }]
    get:
      tags: [Referrals]
      summary: The listener sheet for a session
      description: |
        **One sheet for the whole session**, not one per household — a listener
        is handed a single sheet and scans it for whoever is in front of them.
        Ordered by surname for that reason; a purged household sorts last.

        Open to a team leader, and **the only place a team leader is given the
        reason for referral**. That is a deliberate exception: a listener is
        having a conversation about what went wrong and cannot have it without
        knowing what went wrong. Everywhere else — the referral list, a referral
        by id, the printed picking sheet — the reason stays with administrators,
        and `Referral.reasonId` is still absent for a team leader.

        Deliberately minimal. **No address, postcode, phone, date of birth or
        anything about the referrer.** This ends up on paper in a hall.

        `answers` is the referral's dynamic answers **whole and unfiltered**.
        The sheet's *Cause Details* is one of them, and **the client extracts
        it**: the client owns the form definition and the server holds none, so
        naming the key here would be a guess rather than a contract.
      x-assumed: >-
        Q26 — who appears on the sheet. Every household holding a place is
        listed — `pending_review`, `active` and `reviewed` alike; cancelled and
        rejected ones are not, on the reasoning that
        they are not coming and that handing a volunteer the name and crisis of
        somebody turned away is the harm to avoid. Nobody has confirmed that.
      responses:
        '200':
          description: The sheet
          content:
            application/json:
              schema:
                type: object
                required: [sessionId, households]
                properties:
                  sessionId: { type: string, format: uuid }
                  households:
                    type: array
                    items: { $ref: '#/components/schemas/ListenerSheetHousehold' }
        '404': { description: No such session }
  /api/v1/referrals:
    get:
      tags: [Referrals]
      summary: List referrals
      description: |
        **A team lead does not receive `reasonId`, `referrerEmail`,
        `referrerPhone` or `reviewComment`.** Why someone is hungry is not
        picking information.

        **A team lead's list never contains rejected referrals**, whatever they
        filter by; asking for `status=rejected` returns an empty list rather
        than an error, because the status is simply not one of theirs. Pending
        ones they do see.
      parameters:
        - { in: query, name: sessionId, schema: { type: string, format: uuid } }
        - in: query
          name: status
          schema:
            type: string
            enum: [pending_review, active, reviewed, rejected, cancelled]
      responses:
        '200':
          description: Referrals
          content:
            application/json:
              schema:
                type: object
                required: [referrals]
                properties:
                  referrals:
                    type: array
                    items: { $ref: '#/components/schemas/Referral' }
  /api/v1/referrals/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      tags: [Referrals]
      summary: One referral
      responses:
        '200':
          description: The referral
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Referral' }
        '404':
          description: >-
            No such referral — or it is rejected and the caller is a team lead,
            who has no business learning that it exists.
    patch:
      tags: [Referrals]
      summary: Amend or move a referral
      description: |
        Admin only, and **narrow on purpose: the session and the answers are
        the only things that change.** Supplying `sessionId` moves it; moving
        into a full session requires `acknowledgeOverCapacity: true`, which is
        the server half of the warning the operator has to accept.

        Everything else a referral holds now stands as the referrer sent it —
        see `ReferralAmend` for what went and why a correction goes into the
        form's "other information" answer instead.

        A cancelled or rejected referral cannot be amended **or moved** — a move
        is an amendment, and nothing reinstates a referral, so relocating a dead
        one only moves a dead record.

        **Neither can a referral on a confirmed session**, whichever direction
        it would go: once a session is confirmed nothing on it changes. Moving a
        household off one is the case worth naming, because it would leave them
        recorded against two sessions and change the figures of a session that
        was already signed off.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/ReferralAmend'
                - type: object
                  properties:
                    sessionId: { type: string, format: uuid }
                    acknowledgeOverCapacity: { type: boolean, default: false }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Referral' }
        '409':
          description: >-
            Target session is full and not acknowledged, the referral is
            cancelled or rejected, or its current session has been confirmed
  /api/v1/referrals/{id}/accept:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Referrals]
      summary: Accept a referral awaiting review
      description: >
        Admin only. Moves `pending_review` to `active`. The referral already
        held its place on the session, so accepting it takes no further
        capacity and cannot fail for a full session.

        **Accepting is not the same as having read it.** This settles the
        question the unrecognised address raised; `POST /referrals/{id}/review`
        is the separate pass that records an administrator having read the
        referral through, and it starts from `active`.

        Send `authoriseReferrer` for the second accept button — accept this one
        and add the referrer's address to the authorised list at the same time.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReferralAccept' }
      responses:
        '200':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Referral' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: >-
            That referral is not awaiting review, or `authoriseReferrer` was
            sent for an address already on the authorised list — in which case
            the referral has **not** been accepted either.
        '422':
          description: >-
            `authoriseReferrer` was sent for a referral with no referrer address
            to authorise.
  /api/v1/referrals/{id}/reject:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Referrals]
      summary: Reject a referral awaiting review
      description: >
        Admin only. Moves `pending_review` to `rejected` and releases its place
        on the session. A team lead can no longer see it at all.
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReferralReview' }
      responses:
        '200':
          description: Rejected
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Referral' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: That referral is not awaiting review }
  /api/v1/referrals/{id}/review:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Referrals]
      summary: Record that an administrator has read a referral
      description: >
        Admin only. Moves `active` to `reviewed`.

        **Every referral is meant to be read**, not only the ones held up by an
        unrecognised address, and this is what makes "which has nobody looked at
        yet?" answerable — list `status=active` for the pile still to do.

        Being reviewed changes nothing else: the household holds its place, it
        is picked, and it appears on the listener sheet exactly as before. The
        only thing it says is that somebody has read it.

        No body. There is no comment here — `reviewComment` belongs to the
        accept/reject decision, and no separate review timestamp is kept.

        A referral still `pending_review` is a `409`: decide it first. So is one
        already reviewed, rejected or cancelled.
      responses:
        '200':
          description: Marked as read
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Referral' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { description: That referral is not waiting to be read }
  /api/v1/referrals/{id}/cancel:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    post:
      tags: [Referrals]
      summary: Cancel a referral
      description: >
        Admin only. Idempotent. Frees its place in the session.

        A **rejected** referral is a `409`, not a cancellation. The two are
        different things that happened, and `reviewComment` is the only record
        of why the charity turned the household away — relabelling the status
        would leave that comment on a referral that no longer says a review took
        place.

        A referral on a **confirmed** session is also a `409`: the session is
        closed to every kind of change, cancellation included.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 500 }
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Referral' }
        '409':
          description: That referral was rejected, or its session has been confirmed

  # ------------------------------------------------------------------- sms --
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        `Authorization: Bearer <accessToken>`. Expires after 15 minutes, or
        when the eight-hour sign-in ends if that comes first; see the refresh
        flow in API.md.
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    SessionId:
      in: path
      name: sessionId
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    NotFound:
      description: No such thing.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [error]
      description: >
        Every failure has this shape. `requestId` also appears in the
        `x-request-id` header — quote it when reporting a problem.
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                [
                  BAD_REQUEST,
                  VALIDATION_FAILED,
                  UNAUTHORIZED,
                  FORBIDDEN,
                  NOT_FOUND,
                  CONFLICT,
                  UNPROCESSABLE,
                  INTERNAL_ERROR,
                ]
            message:
              type: string
              description: Safe to show a user. Never contains personal data.
            details:
              type: object
              description: Structured context, such as which fields failed.
            requestId: { type: string }
...
    ReferralAmend:
      type: object
      minProperties: 1
      description: >
        **The answers, and nothing else.** Admin only — there is no
        self-service amendment; a referrer who needs a change phones the food
        bank, and an administrator writes the correction into the form's "other
        information" answer.

        The fixed fields — the names, date of birth, address, postcode, phones,
        household counts, delivery and fuel flags, and the reason — are no
        longer amendable. A referral records what the referrer asked for, and a
        correction typed into the form is read by the people who act on it: the
        answers appear beside the parcel on the picking screen and on the
        listener sheet, which is where somebody acting on a corrected address is
        standing. **A previously-shipped client sending any other field now gets
        a `400`.**

        Which key counts as "other information" belongs to the form, which is
        the client's — so the server takes the answers as a set and does not
        police which of them changed.

        `referrerEmail` is absent for a second reason on top of that: it is what
        the authorisation decision was made on, so editing it would leave a
        referral whose status no longer follows from its address.
      properties:
        answers: { type: object, additionalProperties: true }
    ReferralReview:
      type: object
      description: >
        The administrator's note on why a referral was let through or turned
        away. One short line; it replaces any earlier note, and there is no
        decision history and no decision timestamp. `Referral.referredAt` is the
        timestamp the charity asked for — when the referrer submitted.
      properties:
        comment: { type: string, minLength: 1, maxLength: 200 }
    ReferralAccept:
      type: object
      description: >
        The accept body. `authoriseReferrer` is the **second accept button**:
        let this referral through *and* put its referrer on the authorised list,
        so the next referral from that address is not held up.

        It adds **the referrer's own address and never the domain**. One person
        the charity has decided to trust is not everybody who works where they
        work, and a domain rule taken from a single referral would quietly
        authorise a whole council.

        `organisationName` is required with it and is **the administrator's to
        type and confirm on the screen** — it is not taken from
        `referrerOrganisation`, which is free text the referrer chose. The
        authorised list is the key organisations are reported on, and free text
        fills it with three spellings of one council.

        A `409` if that address is already on the list, in which case nothing
        has happened: the referral is **not** accepted either, so plain accept
        is the next step.
      properties:
        comment: { type: string, minLength: 1, maxLength: 200 }
        authoriseReferrer:
          type: object
          required: [organisationName]
          properties:
            organisationName: { type: string, minLength: 1, maxLength: 200 }
...
    Referral:
      type: object
      description: |
        The staff view. **The last four fields are admin-only** — a team lead
        receives the object without them, so treat them as optional.

        **A team lead does not see rejected referrals at all.** They are absent
        from the list, and fetching one by id is a `404`. Pending ones they do
        see, marked by `status`, because a household awaiting review may well
        turn up and the team lead is the person in the hall when they do.
      # The referee fields and piiPurgedAt are always emitted (they are in the
      # mapper's `base` object, nullable rather than absent). Only the four
      # admin-only fields below are conditional, so only those stay optional.
      required:
        [
          id,
          sessionId,
          status,
          referredAt,
          adults,
          children,
          householdSize,
          isDelivery,
          needsFuelHelp,
          referrerOrganisation,
          referrerName,
          refereeFirstName,
          refereeSurname,
          refereeDateOfBirth,
          refereeAddress,
          refereePostcode,
          refereePhone,
          answers,
          piiPurgedAt,
        ]
      properties:
        id: { type: string, format: uuid }
        sessionId: { type: string, format: uuid }
        status:
          type: string
          enum: [pending_review, active, reviewed, rejected, cancelled]
          description: >
            One pipeline: `pending_review → active → reviewed`, with `rejected`
            off the first step and `cancelled` reachable from any of them.

            `pending_review` — the referrer's address was not recognised and an
            administrator has not yet decided it. It **holds its place** on the
            session, so it counts against capacity and receives a parcel on the
            pick list and printed sheet.

            `active` — it is coming, and **nobody has read it through yet**. A
            referral from a recognised address starts here.

            `reviewed` — an administrator has read it. Identical to `active` in
            every other respect: it holds its place, it is picked, it is on the
            listener sheet. **Filter on `active` and `reviewed` together
            wherever you mean "coming"** — treating `active` alone as the live
            set takes read households off the session.

            `rejected` — an administrator turned it away. Its place is released.

            `cancelled` — withdrawn. A referral cancelled after being read reads
            as `cancelled` and no longer records that anybody read it; the
            charity accepted that when it chose one column over two.
        referredAt:
          type: string
          format: date-time
          description: >-
            When the referrer submitted. There is no separate review timestamp.
        adults: { type: integer }
        children: { type: integer }
        householdSize: { type: integer }
        isDelivery: { type: boolean }
        needsFuelHelp: { type: boolean }
        referrerOrganisation:
          type: string
          description: >-
            As the referrer gave it. For an unrecognised referrer there was no
            authorised-referrer row to derive it from.
        referrerName: { type: [string, 'null'] }
        refereeFirstName: { type: [string, 'null'] }
        refereeSurname: { type: [string, 'null'] }
        refereeDateOfBirth: { type: [string, 'null'], format: date }
        refereeAddress: { type: [string, 'null'] }
        refereePostcode: { type: [string, 'null'] }
        refereePhone: { type: [string, 'null'] }
        answers:
          type: object
          description: >
            Exactly what was submitted. Empty once `piiPurgedAt` is set — see
            below.
          additionalProperties: true
        piiPurgedAt:
          type: [string, 'null']
          format: date-time
          description: >
            When set, the **referee's** fields above are null by design and
            `answers` is empty.

            The referrer's own details are **not** purged — `referrerName`,
            `referrerEmail`, `referrerPhone` and `reviewComment` all survive.
            The retention period exists to forget the household that needed
            feeding, not the professional who referred them.
          x-assumed: >-
            Q12 — that a purge drops the dynamic answers whole. The server no
            longer holds a form definition, so it cannot tell a personal answer
            from a harmless one, and an answer it cannot classify has to be
            assumed personal. Whether the charity needs any answers to survive
            for reporting is still open.
        reasonId:
          type: string
          format: uuid
          description: '**Admin only.** Absent for a team lead.'
        referrerEmail:
          type: [string, 'null']
          description: '**Admin only.**'
        referrerPhone:
          type: [string, 'null']
          description: '**Admin only.**'
        reviewComment:
          type: [string, 'null']
          description: >-
            **Admin only.** The reviewing administrator's one line. It can name a
            referrer or record a suspicion, which is not a team lead's business.
            Who reviewed is stored but never returned — no screen asks.
...
    ListenerSheetHousehold:
      type: object
      description: >-
        One household on the listener sheet. The narrowest response in the API,
        and the only one carrying the reason for referral to a team leader.
      required: [referralId, refereeFirstName, refereeSurname, reason, needsFuelHelp, answers]
      properties:
        referralId: { type: string, format: uuid }
        refereeFirstName: { type: [string, 'null'] }
        refereeSurname: { type: [string, 'null'] }
        reason:
          type: [string, 'null']
          description: >-
            The reason's **label**, not its id. It **survives a purge** — the
            reason is outside the PII block so reporting still works once
            nobody is identifiable. A reason the charity has since retired
            still appears, because the referral was made under it.
        needsFuelHelp: { type: boolean }
        answers:
          type: object
          additionalProperties: true
          description: >-
            The dynamic answers, whole. Extract *Cause Details* here; the server
            does not know which key it is. `{}` once purged.
...